requirements.txt
python-docx==0.8.7
alabaster==0.7.10
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.6.3
Babel==2.5.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
bleach==2.1.3
boto==2.48.0
boto3==1.7.6
botocore==1.10.6
bz2file==0.98
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
cloudpickle==0.5.2
colorama==0.3.9
configparser==3.5.0
cryptography==2.0.3
cycler==0.10.0
decorator==4.3.0
Django==1.9
django-braces==1.11.0
django-crispy-forms==1.6.1
django-formtools==2.0
django-highcharts==0.1.7
django-mysql==2.2.2
django-pure-pagination==0.3.0
django-reversion==2.0.8
django-simple-captcha==0.4.6
django-simple-serializer==2.0.7
docutils==0.14
dwebsocket==0.4.2
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
gensim==3.4.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
image==1.5.24
imagesize==1.0.0
ipaddress==1.0.18
ipykernel==4.8.2
ipython==5.6.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mysql==0.0.1
MySQL-python==1.2.5
nbconvert==5.3.1
nbformat==4.4.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.2
paramiko==2.2.1
parso==0.2.0
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.15
psutil==5.4.5
psycopg2==2.7.4
ptyprocess==0.5.2
pyasn1==0.3.4
pycodestyle==2.4.0
pycparser==2.18
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
PyMySQL==0.8.1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2017.2
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
rope==0.10.7
s3transfer==0.1.13
scandir==1.7
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.7
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
win-unicode-console==0.5
wrapt==1.10.11
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.2.0
